Ignore loading vessels with nan orbits received from srv
