+++ Client/VesselUtilities/VesselCommon.cs
@@ -213,7 +213,7 @@ namespace LunaClient.VesselUtilities
                 owner = LockSystem.LockQuery.GetUpdateLockOwner(vesselId);
             }
 
-            return !string.IsNullOrEmpty(owner) && SystemsContainer.Get<WarpSystem>().PlayerIsInPastSubspace(owner);
+            return !String.IsNullOrEmpty(owner) && SystemsContainer.Get<WarpSystem>().PlayerIsInPastSubspace(owner);
         }
 
         /// <summary>
@@ -269,7 +269,40 @@ namespace LunaClient.VesselUtilities
             return runwayDistance < SettingsSystem.ServerSettings.SafetyBubbleDistance;
         }
 
+        /// <summary>
+        /// Checks if the given config node from a protovessel has NaN orbits
+        /// </summary>
+        public static bool VesselHasNaNPosition(ConfigNode vesselNode)
+        {
+            if (vesselNode.GetValue("landed") == "True" || vesselNode.GetValue("splashed") == "True")
+            {
+                if (Double.TryParse(vesselNode.values.GetValue("lat"), out var latitude)
+                    && (Double.IsNaN(latitude) || Double.IsInfinity(latitude)))
+                    return true;
 
+                if (Double.TryParse(vesselNode.values.GetValue("lon"), out var longitude)
+                    && (Double.IsNaN(longitude) || Double.IsInfinity(longitude)))
+                    return true;
+
+                if (Double.TryParse(vesselNode.values.GetValue("alt"), out var altitude)
+                    && (Double.IsNaN(altitude) || Double.IsInfinity(altitude)))
+                    return true;
+            }
+            else
+            {
+                var orbitNode = vesselNode.GetNode("ORBIT");
+                if (orbitNode != null)
+                {
+                    var allValuesAre0 = orbitNode.values.DistinctNames().Select(v => orbitNode.GetValue(v)).Take(7)
+                        .All(v => v == "0");
+
+                    return allValuesAre0 || orbitNode.values.DistinctNames().Select(v => orbitNode.GetValue(v))
+                               .Any(val => Double.TryParse(val, out var value) && (Double.IsNaN(value) || Double.IsInfinity(value)));
+                }
+            }
+
+            return false;
+        }
 
         /// <summary>
         /// Creates a protovessel from a ConfigNode
